Add two necessary functions to extension.js
We need to add a method to intermittently store both themes and extensions manifests in seperate arrays in session storage(within extension.js). As well as implement a check in the index.js's fetch of theme and extensions (wherever the fetch happens) to replace the data that is typically returned with said fetch with the data we have already cached.