Added snippet: Circular shadow fix for album art
Circular shadow fix for album art is a fix for the square album art shadow you get when you have a circular album art snippet installed. This usually doesn't happen, but some extensions re-enable the shadows in a way that it displays like that. This snippet forcibly turns it circular again.